sr(api): move TrainingTrigger model in devcomp
