+++ api/src/devcomp/domain/models/TrainingTrigger.js
