+++ api/tests/devcomp/integration/infrastructure/repositories/training-repository_test.js
@@ -14,7 +14,7 @@ import { Training } from '../../../../../lib/domain/models/Training.js';
 import { UserRecommendedTraining } from '../../../../../lib/domain/read-models/UserRecommendedTraining.js';
 import { TrainingTriggerForAdmin } from '../../../../../src/devcomp/domain/read-models/TrainingTriggerForAdmin.js';
 import { TrainingForAdmin } from '../../../../../src/devcomp/domain/read-models/TrainingForAdmin.js';
-import { TrainingTrigger } from '../../../../../lib/domain/models/TrainingTrigger.js';
+import { TrainingTrigger } from '../../../../../src/devcomp/domain/models/TrainingTrigger.js';
 import { TrainingTriggerTube } from '../../../../../lib/domain/models/TrainingTriggerTube.js';
 import _ from 'lodash';
